Added match finder so implement traversers logic.

+++ src/main/java/net/jemzart/mutants/domain/dna/DNA.java
@@ -23,7 +23,7 @@ public class DNA {
 	private void validate(String[] raw){
 		for (String str : raw){
 			if (str.length() != length) throw InvalidDNAException.create("DNA must be a square matrix", raw);
-			for (short i = 0; i < length; i++) {
+			for (int i = 0; i < length; i++) {
 				char character = str.charAt(i);
 				if (!isDnaNitrogenousBase(character))
 					throw InvalidDNAException.create("'" + character + "' must be a DNA nitrogenous base", raw);
